rename json reporter file to align with other reporters
